python/scamplers-auth/pyproject.toml
[project]
name = "scamplers-auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Ahmed Said", email = "ahmed.said@jax.org" }]
requires-python = ">=3.13"
dependencies = [
    "hypercorn>=0.17.3",
    "ms-identity-python[quart]",
    "pydantic-settings>=2.8.1",
    "quart>=0.20.0",
]

[tool.uv.sources]
ms-identity-python = { git = "https://github.com/azure-samples/ms-identity-python", rev = "0.9" }
